feat: support versionable states
Events tend to run in parallel and this can trigger race conditions
where one event can overwrite the results of another. To address this,
Proact now supports versionable states able to reconcile different
versions into one.

This commit also ties the type of a Proact Component with the type of
its properties and exposes an API to focus Event Handlers.